move checkbox to subcomponents for the MultiSelect
