+++ packages/components/src/MultiSelect/_docs/Checkbox.stories.module.scss
