+++ packages/components/src/MultiSelect/subcomponents/Checkbox/Checkbox.module.scss
