+++ packages/components/src/MultiSelect/subcomponents/Checkbox/index.ts
@@ -1,0 +1,1 @@
+export * from "./Checkbox"
