+++ packages/components/src/index.ts
@@ -15,4 +15,4 @@ export * from "./InputSearch"
 export * from "./KaizenProvider"
 export * from "./Workflow"
 export * from "./Icons"
-export * from "./Checkbox"
+export * from "./MultiSelect"
